cleaned up warnings and fixed bug in data printing

+++ include/SystemConfig.h
@@ -189,9 +189,9 @@ class ConfigParam {
             char buffer[256];
             sprintf(buffer,"%-18s [%7d,%7d,%7d] %s",
                 name,
-                minVal,
-                val,
-                maxVal,
+                (int)minVal,
+                (int)val,
+                (int)maxVal,
                 desc
             );
             ui->println(buffer);
